+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+##### Bugfixes
+- Fix crash when starting sodalite and starting entry is not a directory
+
 ## v0.14.2: June 26, 2018
 ##### Misc
 - Improve desktop entry